pyproject.toml
[tool.poetry]
name = "reaspeech"
version = "1.0.0"
description = "Speech recognition for REAPER"
homepage  = "https://github.com/TeamAudio/reaspeech/"
license = "https://github.com/TeamAudio/reaspeech/blob/main/LICENSE"
authors = [
    "Dave Benjamin",
    "Mike DeFreitas",
    "Roel Sanchez",
]
readme = "README.md"
packages = [{ include = "app" }]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.10"
unidecode = "^1.3.4"
uvicorn = { extras = ["standard"], version = "^0.18.2" }
gunicorn = "^20.1.0"
tqdm = "^4.64.1"
python-multipart = "^0.0.5"
ffmpeg-python = "^0.2.0"
fastapi = "^0.95.1"
llvmlite = "^0.39.1"
numba = "^0.56.4"
openai-whisper = "20230918"
faster-whisper = "^1.0.2"
torch = [
  {markers = "sys_platform == 'darwin' and platform_machine == 'arm64'", url = "https://download.pytorch.org/whl/cpu/torch-1.13.1-cp310-none-macosx_11_0_arm64.whl"},
  {markers = "sys_platform == 'linux' and platform_machine == 'arm64'", url="https://download.pytorch.org/whl/cpu/torch-1.13.1-cp310-none-macosx_11_0_arm64.whl"},
  {markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-1.13.1-cp310-none-macosx_10_9_x86_64.whl"},
  {markers = "sys_platform == 'linux' and platform_machine == 'aarch64'", url="https://download.pytorch.org/whl/torch-1.13.1-cp310-cp310-manylinux2014_aarch64.whl"},
  {markers = "sys_platform == 'linux' and platform_machine == 'x86_64'", url="https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp310-cp310-linux_x86_64.whl"},
  {markers = "sys_platform == 'win' and platform_machine == 'amd64'", url="https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp310-cp310-win_amd64.whl"},
]
jinja2 = "^3.1.2"
celery = "^5.2.7"
redis = "^4.5.4"
ctranslate2 = "4.2.1"
aiofiles = "^23.2.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
